+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Add a `ConvertableToMimeAddress` interface and trait (#1092)
 
 ### Changed
+- Improve DB-related error message in the testing framework (#1159)
 - Switch to the TYPO3 testing framework (#1130, #1132, #1148)
 
 ### Deprecated